% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm_full.R
\name{cfm_prediction_xy}
\alias{cfm_prediction_xy}
\title{Full Conformal All Scores}
\usage{
cfm_prediction_xy(vec_y, ..., f_pred = cfm_prediction_xy_single, n_core = 1)
}
\description{
All Scores
}
\details{
@export
}
