% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm_full.R
\name{cfm_prediction_xy_single}
\alias{cfm_prediction_xy_single}
\title{Full Conformal for Single X and Y}
\usage{
cfm_prediction_xy_single(
  y,
  x,
  dat,
  fml = "y ~ x",
  alpha = 0.05,
  f_score = cfm_score_abs
)
}
\description{
Full Conformal for Single X and Y
}
