% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfm_package.R
\docType{package}
\name{conformal-package}
\alias{conformal-package}
\alias{conformal}
\alias{#}
\alias{@useDynLib}
\alias{conformal,}
\alias{.registration}
\alias{=}
\alias{TRUE}
\title{The 'conformal' package.}
