\name{LogSDescsTest}
\alias{LogSDescsTest}
\docType{data}
\title{Small Molecule Solubility (LogS) Data: PaDEL descriptors for the test set}
\description{
	PaDEL descriptors have been computed for the test set.
}
\details{

LogSDescsTest is a data frame with PaDEL descriptors for the datapoints in the test set (30\% of the \code{\link{LogS}} data).

}
\seealso{
\code{\link{LogS}}
}
\references{
Wang et al. J. Chem. Inf. Model., 2007, 47 (4), pp 1395-1404
DOI: 10.1021/ci700096r
\url{http://pubs.acs.org/doi/abs/10.1021/ci700096r}
}
