\name{LogSTest}
\alias{LogSTest}
\docType{data}
\title{Small Molecule Solubility (LogS) Data: solubility values for the datapoints in the test set}
\description{
Solubility values for the datapoints in the test set.
}
\details{

LogSTest is a vector containing the solubility values for the datapoints in the test set (30\% of the \code{\link{LogS}} data).

}
\seealso{
\code{\link{LogS}}
}
\references{
Wang et al. J. Chem. Inf. Model., 2007, 47 (4), pp 1395-1404
DOI: 10.1021/ci700096r
\url{http://pubs.acs.org/doi/abs/10.1021/ci700096r}
}
